add more unit test case for rocksdb and rocksdbsst
