+++ hugegraph-test/src/main/java/com/baidu/hugegraph/unit/UnitTestSuite.java
@@ -55,6 +55,7 @@ import com.baidu.hugegraph.unit.id.SplicingIdGeneratorTest;
 import com.baidu.hugegraph.unit.mysql.MysqlUtilTest;
 import com.baidu.hugegraph.unit.mysql.WhereBuilderTest;
 import com.baidu.hugegraph.unit.rocksdb.RocksDBCountersTest;
+import com.baidu.hugegraph.unit.rocksdb.RocksDBSessionTest;
 import com.baidu.hugegraph.unit.rocksdb.RocksDBSessionsTest;
 import com.baidu.hugegraph.unit.serializer.BinaryBackendEntryTest;
 import com.baidu.hugegraph.unit.serializer.BinaryScatterSerializerTest;
@@ -130,6 +131,7 @@ import com.baidu.hugegraph.unit.util.VersionTest;
 
     /* rocksdb */
     RocksDBSessionsTest.class,
+    RocksDBSessionTest.class,
     RocksDBCountersTest.class,
 
     /* utils */
